Added pgaussian
