+++ src/TMB/pgaussian.hpp
@@ -1,0 +1,20 @@
+#include "LocalCop/gaussian.hpp"
+
+#undef TMB_OBJECTIVE_PTR
+#define TMB_OBJECTIVE_PTR obj
+
+template <class Type>
+Type pgaussian(objective_function<Type> *obj) {
+  // R inputs
+  DATA_VECTOR(u1);
+  DATA_VECTOR(u2);
+  DATA_VECTOR(weights);
+  PARAMETER_VECTOR(theta);
+  // output
+  vector<Type> lcdf = LocalCop::pgaussian(u1, u2, theta, 1);
+  lcdf.array() *= weights.array();
+  return -lcdf.sum();
+}
+
+#undef TMB_OBJECTIVE_PTR
+#define TMB_OBJECTIVE_PTR this
